whisper/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
torchaudio==2.3.1
openai-whisper==20231117
soundfile==0.12.1
